# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope:
    properties:
      data: {}
      details: {}
      errorCode:
        type: string
      errorMsg:
        type: string
      requestId:
        type: string
    type: object
...
  internal_transport_http_usersettings.PatchSettingsRequest:
    properties:
      settings:
        additionalProperties:
          type: string
        type: object
    required:
    - settings
    type: object
  internal_transport_http_usersettings.UserSettingsResponse:
    properties:
      settings:
        additionalProperties:
          type: string
        type: object
    type: object
  internal_transport_http_usersettings.UserSettingsResponseDoc:
    properties:
      data:
        $ref: '#/definitions/internal_transport_http_usersettings.UserSettingsResponse'
      errorMsg:
        type: string
    type: object
info:
  contact: {}
  description: DEEIX Chat 后端 API 文档
  title: DEEIX Chat API
  version: "0.3.0"
paths:
...
  /user/settings:
    get:
      description: 返回当前用户全部个人偏好配置，缺失项以默认值填充
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_usersettings.UserSettingsResponseDoc'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 获取当前用户的配置
      tags:
      - user/settings
    patch:
      consumes:
      - application/json
      description: 批量更新用户个人偏好配置，返回更新后的全量配置
      parameters:
      - description: 更新项
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_transport_http_usersettings.PatchSettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_http_usersettings.UserSettingsResponseDoc'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_DEEIX-AI_DEEIX-Chat_backend_internal_shared_response.Envelope'
      security:
      - BearerAuth: []
      summary: 更新当前用户的配置
      tags:
      - user/settings
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"